+++ common/frame.h
@@ -41,16 +41,16 @@ typedef struct x264_frame
     int64_t i_pts;
     int64_t i_dts;
     int64_t i_reordered_pts;
-    int     i_duration;  /* in SPS time_scale units (i.e 2 * timebase units) used for vfr */
+    int64_t i_duration;  /* in SPS time_scale units (i.e 2 * timebase units) used for vfr */
     float   f_duration;  /* in seconds */
-    int     i_cpb_duration;
-    int     i_cpb_delay; /* in SPS time_scale units (i.e 2 * timebase units) */
-    int     i_dpb_output_delay;
+    int64_t i_cpb_duration;
+    int64_t i_cpb_delay; /* in SPS time_scale units (i.e 2 * timebase units) */
+    int64_t i_dpb_output_delay;
     x264_param_t *param;
 
     int     i_frame;     /* Presentation frame number */
     int     i_coded;     /* Coded frame number */
-    int     i_field_cnt; /* Presentation field count */
+    int64_t i_field_cnt; /* Presentation field count */
     int     i_frame_num; /* 7.4.3 frame_num */
     int     b_kept_as_ref;
     int     i_pic_struct;
@@ -134,8 +134,8 @@ typedef struct x264_frame
     uint8_t i_planned_type[X264_LOOKAHEAD_MAX+1];
     int i_planned_satd[X264_LOOKAHEAD_MAX+1];
     double f_planned_cpb_duration[X264_LOOKAHEAD_MAX+1];
-    int i_coded_fields_lookahead;
-    int i_cpb_delay_lookahead;
+    int64_t i_coded_fields_lookahead;
+    int64_t i_cpb_delay_lookahead;
 
     /* threading */
     int     i_lines_completed; /* in pixels */
